Prepare HTML renderer for release
