--- benchmarks/.html-v2-parity-trigger
@@ -1,1 +1,0 @@
-plain-html-parity
